+++ .github/workflows/e2e-subtensor-tests.yml
@@ -53,8 +53,8 @@ jobs:
 
       - name: Add wasm32-unknown-unknown target
         run: |
-          rustup target add wasm32-unknown-unknown
-          rustup component add rust-src
+          rustup target add wasm32-unknown-unknown --toolchain stable-x86_64-unknown-linux-gnu
+          rustup component add rust-src --toolchain stable-x86_64-unknown-linux-gnu
 
       - name: Clone subtensor repo
         run: git clone https://github.com/opentensor/subtensor.git
